[Android/Feat] 워치 스피커 api 연동

See merge request s12-final/S12P31D103!130

+++ Android/app/src/main/java/com/hogumiwarts/lumos/ui/screens/control/light/LightIntent.kt
@@ -8,7 +8,7 @@ sealed class LightIntent {
 
     data class ChangeLightBright(val deviceId: Long, val brightness: Int): LightIntent()
 
-    data class ChangeLightColor(val deviceId: Long, val color: Int, val saturation:Float=100f): LightIntent()
+    data class ChangeLightColor(val deviceId: Long, val color: Float, val saturation:Float=100f): LightIntent()
 
     data class ChangeLightTemperature(val deviceId: Long, val temperature: Int): LightIntent()
