+++ Android/data/src/main/java/com/hogumiwarts/data/entity/remote/Response/light/LightColorResponse.kt
@@ -2,6 +2,6 @@ package com.hogumiwarts.data.entity.remote.Response.light
 
 data class LightColorResponse(
     val success: Boolean,
-    val hue: Int,
+    val hue: Float,
     val saturation: Float
 )
