+++ Android/domain/src/main/java/com/hogumiwarts/domain/model/light/LightColorData.kt
@@ -2,6 +2,6 @@ package com.hogumiwarts.domain.model.light
 
 data class LightColorData(
     val success : Boolean,
-    val hue: Int,
+    val hue: Float,
     val saturation: Float,
 )
